Refine models
Make className a first param for the define method.
Change how property validations are defined.
Before I used the Rails style but that would not allow for automatic UI element restrictions.
For instance setting the maximum length on a HTML input element based on the validation property.
Now it could be bound to "model.clazz.properties.name.maximumLength"